Fix heat-3d border issues in order to match behaviour of polybench
In polybench, the shell of the full cube is not updated. As a result,
it is nescessary to check these boundary condition when computing the
all the subproblems.

If a given cube doesn't have a neighbor in a given direction, that face
shall not be updated. However care must also be taken on faces that do
have a neighbor, but are next to faces that don't.

For example a cube with a TOP neighbor, but no LEFT neighbor. The TOP
face should be updated, but the LEFT face and the TOP-LEFT edge shouldn't
be updated!

Extra care must be taken for the corners. These are only updated if all
three neighbors are present.